examples/chatgpt/rag-quickstart/gcp/requirements.txt
google-cloud-bigquery
functions-framework
openai